# openapi.yaml
openapi: 3.0.0
info:
  title: Firestore CRUD API
  description: An API for performing CRUD operations on users.
  version: 1.0.0
servers:
  - url: https://users-164502969077.asia-southeast1.run.app
paths:
  /users:
    put:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                uid:
                  type: string
      responses:
        '201':
          description: User created successfully
        '200':
          description: Document already exists.
        '400':
          description: Name, email, and uid are required
        '500':
          description: Error creating user
  /users/{uid}:
    get:
      summary: Read a user by UID
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User retrieved successfully
        '400':
          description: UID is required
        '404':
          description: User not found
        '500':
          description: Error reading user
    patch:
      summary: Update a user by UID
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                qualification:
                  type: string
                profession:
                  type: string
      responses:
        '200':
          description: User updated successfully
        '400':
          description: At least one field to update is required
        '404':
          description: User not found
        '500':
          description: Error updating user
    delete:
      summary: Delete a user by UID
      parameters:
        - name: uid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
        '400':
          description: UID is required
        '404':
          description: User not found
        '500':
          description: Error deleting user
  /users/slug/{slugURL}:
    get:
      summary: Get a user by slugURL
      parameters:
        - name: slugURL
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User retrieved successfully
        '404':
          description: User not found
        '500':
          description: Error getting user by slugURL